Restructured project, added 'test data', restructured main.py, fixed imports.
