+++ .gitignore
@@ -1,2 +1,11 @@
-*/__pycache__
+# Default ignored files
+/shelf/
+/workspace.xml
+# Editor-based HTTP Client requests
+/httpRequests/
+# Datasource local storage ignored files
+/dataSources/
+/dataSources.local.xml
+# Pycache
 /__pycache__
+*/__pycache__
